database/pyproject.toml
[tool.poetry]
authors = ["Daven Quinn <dev@davenquinn.com>"]
description = "A SQLAlchemy-based database toolkit."
name = "macrostrat.database"
packages = [{ include = "macrostrat" }]
version = "3.3.0"

[tool.poetry.dependencies]
GeoAlchemy2 = "^0.14.0"
SQLAlchemy = "^2.0.18"
SQLAlchemy-Utils = "^0.41.1"
click = "^8.1.3"
"macrostrat.utils" = "^1.0.0"
psycopg2-binary = "^2.9.6"
python = "^3.8"
sqlparse = "^0.4.4"
aiofiles = "^23.2.1"
rich = "^13.7.1"

[tool.poetry.dev-dependencies]
"macrostrat.utils" = { path = "../utils", develop = true }

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
